lyra: properly call time()
time() actually has an outptr argument, and we don't control rdi on
ifunc invocation. The previous code works by luck on Ubuntu 2204 and
Debian testing, but segfault in the backdoor injection code completely
defeats the goal of this challenge (hiding).

We addressed that by calling time() in the backdoor injection code. This
is less obscure, but seems to be the best thing we could do, as there
are no no-argument func for fetching current wall clock time.

The change also increases the backdoor injection code size by 13 bytes
(which is a lot), which hopefully could be hammered down in further
changes.